Inline ReactErrorUtils
This is pretty simple now.
