+++ packages/react-native-renderer/src/legacy-events/EventBatching.js
@@ -6,12 +6,10 @@
  * @flow
  */
 
-import {rethrowCaughtError} from 'shared/ReactErrorUtils';
-
 import type {ReactSyntheticEvent} from './ReactSyntheticEventType';
 import accumulateInto from './accumulateInto';
 import forEachAccumulated from './forEachAccumulated';
-import {executeDispatchesInOrder} from './EventPluginUtils';
+import {executeDispatchesInOrder, rethrowCaughtError} from './EventPluginUtils';
 
 /**
  * Internal queue of events that have accumulated their dispatches and are
